RENAME : VITE 로 마이그레이션에 따른 파일 확장자 변경
